+++ src/client/order/CategoryPanel.java
@@ -1,6 +1,9 @@
 package client.order;
 
 import client.MainFrame;
+import org.apache.ibatis.session.SqlSession;
+import org.apache.ibatis.session.SqlSessionFactory;
+
 import javax.swing.*;
 import java.awt.*;
 import java.util.List;
@@ -9,20 +12,19 @@ public class CategoryPanel extends JPanel {
 
     private final MenuPanel menuPanel;
     private String currentCategory;
+    SqlSessionFactory factory;
 
     public CategoryPanel(MenuPanel menuPanel, MainFrame mainFrame) {
         this.menuPanel = menuPanel;
         super.setLayout(new BorderLayout());
         super.setBackground(Color.WHITE);
-
+        this.factory= mainFrame.factory;
         this.currentCategory = "모든 메뉴";
 
         JPanel buttonContainerPanel = new JPanel(new GridLayout(0,4, 10, 10));
         buttonContainerPanel.setBackground(Color.WHITE);
 
         buttonContainerPanel.setBorder(BorderFactory.createEmptyBorder(5, 0, 5, 0));
-
-        ProductsDao productsDao = new ProductsDao(mainFrame.factory);
 
         JButton allButton = new JButton("모든 메뉴");
         allButton.setFont(new Font("맑은 고딕", Font.BOLD, 14));
@@ -34,7 +36,7 @@ public class CategoryPanel extends JPanel {
         });
         buttonContainerPanel.add(allButton);
 
-        List<String> categories = productsDao.getCategories();
+        List<String> categories = getCategories();
         if (categories != null && !categories.isEmpty()) {
             for (String categoryName : categories) {
                 JButton categoryButton = new JButton(categoryName);
@@ -59,5 +61,14 @@ public class CategoryPanel extends JPanel {
         scrollPane.getHorizontalScrollBar().setUnitIncrement(16);
 
         super.add(scrollPane, BorderLayout.CENTER);
+    }
+
+    public List<String> getCategories() {
+        try (SqlSession session = factory.openSession()) {
+            return session.selectList("products.getCategories");
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 }
